Reduce needless garbage generation.
The calls to msgf() show up as a garbage creation hotspot in go 1.7
due to the creation of the []interface to hold 'args'. The compiler
wasn't smart enough to elide that away (of it is required for some
reason I cannot fathom). So switch to passing in pointer to a known
type as a single arg. That happens to work for the format msgs we
used, and the compiler does much better with that.