remove the goroutine blocking/bouncing when constructing a syncrequest
now instead of asking each consumer.run serially to add itself to the
OffsetCommitRequest, we ask all concurrently to report to us what
to add, and we (client.run) add it. This, along with a correctly
sized channel, removes the stuttering visible in goroutine trace
and reduces the time the client.run stays busy. (though that last
bit isn't that important, since the main delay is in waiting for
the response to the commit, which depends on network RTT and kafka
broker internal latency.