a stable and consistent partitioner. WIP. Compiles
and ought to be code complete if I haven't made a mistake, but
lacks any unit tests.

This partitioner is stable in the sense that when group membership
changes it reassigns the minimum number of partitions.

This partitioner is consistent in the sense that two topics with
identical group memberships and an identical # of partitions
receive the same partitioning.

That allows you to have messages with the same key but on different
topics arrive at the same consumer. That is perfect if the consumer
is holding state related to the key.