clean up some edge cases in adjust_partitioning()
so that if we have a lot of members with 'high' we take from
those to give to those with less with 'low'. In the end all
members have low<=n<=high partitions assigned to each, and
either low==high (if the # of members divides the # of partitions)
or low==high-1 if there must be some uneveness

This code works. I suspect it is doing one more pass than
necessary, since the truncation to [:high] is identical to
stealing_from_the_numerous later on, but I'll develop tests
with this code and then rethink it.