react to an increase in partition count by rejoining
This way when we scale up partitions on a live
cluster the consumers start consuming from the
new partitions without needing to be restarted.

Note that kafka (0.9.1) does not bump consumer
group generations when a topic's partition count changes.
(thinking about it, it can't. the broker doesn't know
how to decode the join/sync messages to retrieve the
topic list in the most general case, though it could
be done for the 'consumer' protocol)

The sarama.Client polls the metadata every
sarama.Config.Metadata.RefreshFrequency. There isn't
a way for us to hook into that. So we poll sarama.Client
at the same interval. That way it takes no more than 1.999
intervals for us to notice a change.

We check that the number of partitions still agrees
with what we saw before we joined the consumer group,
and rejoin if it differs.

Note that the sarama default for Metadata.RefreshFrequency
is 10 minutes, which can be a long time to wait. If it is
too long I suggest you set Metadata.RefreshFrequency to
something smaller.