move sidechannel send code into sidechannel goroutine[1]
so I can sent to sidechannel when cleanly abandoning a partition.
(We also end up writing when we loose the partition due to a rebalance,
but in that case it's a race between the new owner getting their
SyncGroupReponse and our write to the sidechannel, and the sidechannel
almost always looses.

[1] though really it could be in its own goroutine. it is independent
of consuming